Fix grouped FP8 rowwise DIVUP type deduction
Cast the rowwise vector width to size_t when computing vectorized grouped FP8 launch blocks so DIVUP deduces a single type without changing grouped quantization semantics.

Orchestra-Work-Order: wo_7a6b82d5824c69e4d455d52bf2db05d5
Orchestra-Task: task_6a951997bb0976222d1ed1201be1a7cf
Orchestra-Run: run_9b8fc0cf9ba802ea8ce8d596a577df12

+++ transformer_engine/common/cast/fp8/group_quantize_fp8.cuh
@@ -466,7 +466,8 @@ void group_quantize_current_scaling(const GroupedTensor *input, const Tensor *no
                   is_aligned_to(output->data.dptr, sizeof(OType) * rowwise_nvec);
 
               if (use_vectorized_rowwise) {
-                const size_t average_tensor_vecs = DIVUP(average_tensor_elements, rowwise_nvec);
+                const size_t average_tensor_vecs =
+                    DIVUP(average_tensor_elements, static_cast<size_t>(rowwise_nvec));
                 const size_t rowwise_blocks_per_tensor = limited_blocks_per_tensor(
                     average_tensor_vecs, ROWWISE_THREADS_PER_BLOCK);
                 const dim3 rowwise_grid(rowwise_blocks_per_tensor, num_tensors);
